nicegui_admin/requirements.txt
nicegui==1.4.10
python-dotenv==1.0.0
requests==2.31.0
